Fix float32 promotions in spatial attention and SSM skip connection
Two remaining AMP memory leaks:
1. spatial_bias (float32) promoted attention scores (B*T,H,N,N) to float32,
   wasting ~1 GB per spatial layer (2 layers = 2 GB)
2. self.D (float32 param) promoted SSM skip connection to float32,
   wasting ~41 MB per block (4 blocks = 164 MB)

Cast both to the running dtype (float16 under AMP) to keep all
activations in half precision. Together saves ~2.3 GB on the
10 GB MIG slice.

+++ model/mamba_temporal.py
@@ -125,7 +125,7 @@ class SelectiveSSM(nn.Module):
         A = -torch.exp(self.A_log)  # (d_inner, d_state)
 
         y = self._ssm_scan(x_conv, delta, A, B_mat, C_mat)
-        y = y + x_conv * self.D  # skip connection
+        y = y + x_conv * self.D.to(y.dtype)  # skip connection (keep float16 under AMP)
         y = y * F.silu(z)  # gating
 
         return self.out_proj(y) + residual
